Regenerate project with cmake-init
This makes the project use modern CMake, spellchecker and clang-format
linting in CI, vcpkg for dependencies and cleans up the tests.

Note that while vcpkg does have a port for nanobench, it does not
properly install a CMake package, so an overlay port had to be written
for that.

The content section has been removed from the README, because GitHub
already implements a Table of Contents feature for READMEs, so this
serves no purpose in the document itself.